requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
idna==3.10
pillow==10.4.0
pycparser==2.22
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
urllib3==2.2.3
